Populate the initial config with known and trusted services
Users should still verify that the config is correct, and compare
the config against additional sources, but at least this would be a good
start and will help users bootstrap their configurations.
This is only set when users install the extension. Existing user
configurations are never modified!

At the moment only EteSync is there, with more to follow.